Increase version 0.7 Fix bugs

+++ qrshell.sh
@@ -1,16 +1,13 @@
 #!/usr/bin/env bash
-
 #-----------------------------------------------------------------------------|
 # AUTOR             : Matheus Martins <3mhenrique@gmail.com>
 # HOMEPAGE          : https://github.com/mateuscomh/qrbash
-# DATE/VER.         : 16/09/2023 0.6
+# DATE/VER.         : 16/09/2023 0.7
 # LICENCE           : GPL3
 # SHORT DESC        : Shell Script to generate and read QRCode from terminal
 # DEPS              : zbarimg(zbar-tools), qrencode, scrot, xclip(optional)
 #-----------------------------------------------------------------------------|
-
 USAGE="Generate and read QR Codes from terminal
-
  ▄▄▄▄▄▄▄▄▄▄▄  ▄▄▄▄▄▄▄▄▄▄▄  ▄▄▄▄▄▄▄▄▄▄▄  ▄         ▄  ▄▄▄▄▄▄▄▄▄▄▄  ▄            ▄           
 ▐░░░░░░░░░░░▌▐░░░░░░░░░░░▌▐░░░░░░░░░░░▌▐░▌       ▐░▌▐░░░░░░░░░░░▌▐░▌          ▐░▌          
 ▐░█▀▀▀▀▀▀▀█░▌▐░█▀▀▀▀▀▀▀█░▌▐░█▀▀▀▀▀▀▀▀▀ ▐░▌       ▐░▌▐░█▀▀▀▀▀▀▀▀▀ ▐░▌          ▐░▌          
@@ -21,9 +18,7 @@ USAGE="Generate and read QR Codes from terminal
 ▐░░░░░░░░░░░▌▐░▌     ▐░▌            ▐░▌▐░▌       ▐░▌▐░▌          ▐░▌          ▐░▌          
  ▀▀▀▀▀▀█░█▀▀ ▐░▌      ▐░▌  ▄▄▄▄▄▄▄▄▄█░▌▐░▌       ▐░▌▐░█▄▄▄▄▄▄▄▄▄ ▐░█▄▄▄▄▄▄▄▄▄ ▐░█▄▄▄▄▄▄▄▄▄ 
         ▐░▌  ▐░▌       ▐░▌▐░░░░░░░░░░░▌▐░▌       ▐░▌▐░░░░░░░░░░░▌▐░░░░░░░░░░░▌▐░░░░░░░░░░░▌
-         ▀    ▀         ▀  ▀▀▀▀▀▀▀▀▀▀▀  ▀         ▀  ▀▀▀▀▀▀▀▀▀▀▀  ▀▀▀▀▀▀▀▀▀▀▀  ▀▀▀▀▀▀▀▀▀▀▀ 
-                                                                                           
-"
+         ▀    ▀         ▀  ▀▀▀▀▀▀▀▀▀▀▀  ▀         ▀  ▀▀▀▀▀▀▀▀▀▀▀  ▀▀▀▀▀▀▀▀▀▀▀  ▀▀▀▀▀▀▀▀▀▀▀ "
 show_help() {
   echo -e "$USAGE"
   echo "Uso: $0 [string]"
@@ -43,7 +38,6 @@ if [ ! -d "$TMPDIR" ]; then
   echo "/dev/shm não está disponível. Usando /tmp como alternativa."
   TMPDIR="/tmp"
 fi
-TMPFILE=$(mktemp "$TMPDIR/screenshot.XXXXXX.png")
 
 # Main
 echo -e "$USAGE"
@@ -51,12 +45,13 @@ if [ ! "$1" ]; then
   cd "$TMPDIR" || exit
   echo "Use mouse select an area on screen to decode QR"
   OUTPUT=$(scrot -s -e 'echo Image size $w x $h ; echo "" ; zbarimg "$f"')
-  echo -e "$OUTPUT \n" | tee /dev/tty | sed -n "s/.*QR-Code://p" | xclip -selection clipboard >/dev/null
+  echo -e "$OUTPUT \n" | tee /dev/tty | sed -n "s/.*QR-Code://p" | xclip -selection clipboard &>/dev/null
 elif [ -f "$1" ]; then
   OUTPUT=$(zbarimg "$1")
-  echo -e "$OUTPUT \n" | tee /dev/tty | sed -n "s/.*QR-Code://p" | xclip -selection clipboard >/dev/null
-elif [ "$#" -eq 1 ]; then
-  qrencode -m 2 -t ANSIUTF8 "$1" | tee "$TMPFILE"
+  echo -e "$OUTPUT \n" | tee /dev/tty | sed -n "s/.*QR-Code://p" | xclip -selection clipboard &>/dev/null
+elif [ "$#" -eq 1 -a ! -f "$1" ]; then
+  TMPFILE=$(mktemp "$TMPDIR/screenshot.XXXXXX.txt") 
+  qrencode -m 2 -t ANSIUTF8 "$1" | tee "$TMPFILE" 
 else
   echo "Invalid input"; show_help
   exit 1
